api/requirements.txt
fastapi
uvicorn
psycopg2-binary
apify-client
google-genai==1.70.0
mangum
python-dotenv
